# 🚀 Don't use regexp for no-cache value check