@kiyonlin 2020-07-26
Nice work @larrylv . It is worth mentioning that we just consider standard http header values until now.